requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.3
cron-descriptor==1.4.5
cryptography==43.0.3
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.1.0
django-celery-beat==2.7.0
django-cors-headers==4.5.0
django-environ==0.11.2
django-timezone-field==7.0
django_mock_queries==2.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
factory_boy==3.3.1
Faker==30.6.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
model-bakery==1.20.0
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
pytest-mock==3.14.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
